+++ docs/package-search-index.js
@@ -1,0 +1,1 @@
+packageSearchIndex = [{"l":"All Packages","url":"allpackages-index.html"},{"l":"de.muspellheim.commons.fx"},{"l":"de.muspellheim.commons.fx.control"},{"l":"de.muspellheim.commons.fx.dialog"},{"l":"de.muspellheim.commons.fx.test"}]