+++ src/main/java/de/muspellheim/commons/fx/dialog/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Muspellheim Commons FX
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+/**
+ * Contains general GUI dialogs like about and exception.
+ */
+package de.muspellheim.commons.fx.dialog;