+++ src/main/java/de/muspellheim/commons/fx/test/package-info.java
@@ -1,0 +1,9 @@
+/*
+ * Muspellheim Commons FX
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+/**
+ * Contains extension for UI tests with JUnit 5.
+ */
+package de.muspellheim.commons.fx.test;