+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.0.0] - 2019-10-31
+
+### Added
+
+*   Date picker control.
+*   View controller factory will load resource bundle if it exists.
+
+### Changed
+
+*   Added sub packages and moved some classes to new packages.
+*   Compensated different generic types order of `StageController` and
+    `ViewControllerFactory`.
+*   Replaced about dialog.
+
 ## [v1.0.0] - 2019-10-27
 
 ### Added
@@ -17,5 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 *   Stage controller based on view controller.
 
 
-[Unreleased]: https://github.com/falkoschumann/java-muspellheim-commons-fx/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/falkoschumann/java-muspellheim-commons-fx/compare/v2.0.0...HEAD
+[v2.0.0]: https://github.com/falkoschumann/java-muspellheim-commons-fx/compare/v1.0.0...v2.0.0
 [v1.0.0]: https://github.com/falkoschumann/java-muspellheim-commons-fx/tree/v1.0.0